backend/requirements.txt
# Django
Django==5.0.10
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# DB
psycopg[binary]==3.2.3
dj-database-url==2.3.0

# 운영
gunicorn==23.0.0
whitenoise==6.8.2
python-decouple==3.8

# 데이터 import
openpyxl==3.1.5

# (인증 스키마 미리 — D18, 라우팅은 비활성)
djangorestframework-simplejwt==5.3.1
